// front/src/shared/api/generated/model/orderCreate.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { OrderCreateTotalAmount } from './orderCreateTotalAmount';
import type { OrderItemCreate } from './orderItemCreate';

export interface OrderCreate {
  full_name: string;
  email: string;
  address: string;
  city: string;
  postal_code: string;
  phone: string;
  total_amount: OrderCreateTotalAmount;
  order_items: OrderItemCreate[];
}