// front/src/shared/api/generated/model/productInDB.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInDBDescription } from './productInDBDescription';
import type { ProductInDBDiscount } from './productInDBDiscount';
import type { ProductInDBCharacteristics } from './productInDBCharacteristics';
import type { ProductInDBImageUrl } from './productInDBImageUrl';
import type { ProductInDBCategory } from './productInDBCategory';
import type { ProductInDBUpdatedAt } from './productInDBUpdatedAt';

export interface ProductInDB {
  name: string;
  slug: string;
  description?: ProductInDBDescription;
  price: string;
  discount?: ProductInDBDiscount;
  characteristics?: ProductInDBCharacteristics;
  imageUrl?: ProductInDBImageUrl;
  categoryId: string;
  id: string;
  category?: ProductInDBCategory;
  createdAt: string;
  updatedAt?: ProductInDBUpdatedAt;
  timesOrdered: number;
  offlinePurchases?: number;
}